add getContacts and getContactsById
