+++ app.js
@@ -2,7 +2,8 @@ const express = require('express');
 const logger = require('morgan');
 const cors = require('cors');
 
-const contactsRouter = require('./src/routes/api/contacts.js');
+const contactsRouter = require('./src/routers/usersRouter');
+
 
 const app = express();
